Replace login view with onboarding
